refactor(file structure): adopt kebab-case for file naming convention across app
- Align file naming with Google's best practices by transitioning from camelCase to kebab-case, enhancing readability and consistency across the application.

- Add slight reformatting for code consistency
