+++ app.js
@@ -1,8 +1,9 @@
 const express = require('express');
 const PORT = process.env.PORT || 3000;
-const connectDB = require('./config/dbConnection');
-const viewRouter = require('./routes/viewRoutes');
-const apiRouter = require('./routes/apiRoutes');
+const connectDB = require('./config/db-connection');
+const viewRouter = require('./routes/view-routes');
+const apiRouter = require('./routes/api-routes');
+const budgetRoutes = require('./routes/budget-routes');
 
 const { initializeApp} = require('firebase/app');
 const { getAuth } = require('firebase/auth');
@@ -37,6 +38,6 @@ app.set('view engine', 'ejs');
 // app.use('/login', loginRouter);
 app.use('/api', apiRouter);
 app.use('/', viewRouter);
-
+app.use('/', budgetRoutes);
 
 app.listen(PORT, () => {console.log(`Server is running on port ${PORT}`);});
