--- views/partials/visitorHeader.ejs
@@ -1,11 +1,0 @@
-<nav class="navbar navbar-expand-lg navbar-light bg-light">
-    <a class="navbar-brand" href="/">App Name</a>
-    <ul class="navbar-nav mr-auto">
-      <li class="nav-item">
-        <a class="nav-link" href="/registration-form">Sign Up</a>
-      </li>
-      <li class="nav-item">
-        <a class="nav-link" href="/login-form">Login</a>
-      </li>
-    </ul>
-  </nav>
